requirements.txt
tensorflow==2.15.0
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
